Add files via upload
One of the things that I noticed within my dataset were several instances where the birth rate, death rate, and fertility rate was 0. After doing extensive research on each country where this was the case, I determined that the most common reason for this was because data for the real rates was not available. As such, I made several SQL queries that set those values to null so they wouldn't be counted later during analysis.