Add files via upload
After cleaning and creating SQL queries, I used the queries to make temp tables which I then uploaded to Tableau to analyze further. The visuals get me closer to answering the questions I had.
In the dashboard, in the top left, we have a visual regarding population growth rates, where the bluer a country is, the more it's growing, and the more orange a country is, the more it's losing population. It appears that the region of the world growing the quickest is Sub-Saharan Africa, with central Asia also becoming more developed. However, Eastern Europe seems to be becoming less populated.
In the center left, there's a visual pertaining to urbanization. Similarly, the bluer a country is, the more of a percentage that's moving to urban areas, and the more orange a country is, the more of a percentage that's leaving cities and heading to rural areas. We can immediately spot a distinct correlation between urbanization rates and population growth. This is because almost all the countries that gained population also had more of its population heading to urban areas, and vice versa. Going off of that, we can observe that urbanization rates are high in Sub-Saharan Africa and low in Eastern Europe, as well as Venezuela, who also experienced population loss. We can generally observe that, with some exceptions, more people are moving to urban areas than rural areas.
In the bottom left, we look into life expectancy changes in terms of years. Orange countries have lower life expectancies than it did in 2017 while blue countries have higher life expectancies. The first thing we notice is that life expectancies are generally lower than they were in 2017. However, there are two areas which defy these trends: Sub-Saharan Africa and East Asia. While East Asia certainly did not have quite as strong urbanization rates or population growth as Sub-Saharan Africa did, it did see more of its people moving to urban areas and their countries growing in population.
In the top right, we observe countries in terms of a birth to death ratio, where bluer countries have significantly higher deaths compared to births than countries in turquoise. While Sub-Saharan Africa is also observing more births compared to deaths, two more regions to consider are the Middle East and Central Asia, which also have more people being born than dying. Eastern Europe seems to be having more people die than those being born, which may be part of the reason why its population is decreasing.
Finally, in the bottom right, we analyze infant mortality rates compared to fertility rates. Bluer countries represent those countries with a large infant mortality rate compared to fertility rate. Bluer countries may observe more births, but may not have the adequate resources to make sure the babies live long, fulfilling lives ahead of them. The two areas that would appear to be the case are Sub-Saharan Africa and Southeast Asia. However, with life expectancies increasing in both regions, along with more urbanization and population growth, it appears that both regions could become very developed in the future if trends continue.
In conclusion, we see that Sub-Saharan Africa and Southeast Asia are becoming more developed areas of the world, while Eastern Europe is becoming less developed than it once was.